+++ playwright.config.ts
@@ -14,7 +14,7 @@ const config: PlaywrightTestConfig = {
   reporter: 'html',
   use: {
     baseURL: 'http://localhost:3000',
-    trace: 'on-first-retry',
+    trace: 'on',
     screenshot: 'only-on-failure',
     video: 'retain-on-failure',
     ignoreHTTPSErrors: true,